create intervals_utils.py to avoid circular imports
